# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API document for FilesBank",
        "version": "1.0.0",
        "title": "FilesBank",
        "termsOfService": "",
        "contact": {
            "email": "gpaitoo@myself.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:5000",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/upload-file": {
            "post": {
                "tags": [
                    "upload-file"
                ],
                "summary": "Upload new file",
                "description": "convert file to base 64 and parse it here",
                "operationId": "upload-file",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Pet object that needs to be added to the store",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Upload"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "upload-multipart"
                ],
                "summary": "Upload multipart file",
                "description": "",
                "operationId": "uploadMultipart",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "file to upload",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Invalid Token Supplied"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        },
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
            "flow": "implicit",
            "scopes": {
                "read:pets": "read your pets",
                "write:pets": "modify pets in your account"
            }
        }
    },
    "definitions": {
...
        "Upload": {
            "type": "object",
            "properties": {
                "FileType": {
                    "type": "string"
                },
                "UploadMode": {
                    "type": "string"
                },
                "File": {
                    "type": "string"
                }
            }
        },
...
    },
...
}